requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.3.0
bitarray==2.9.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Cython==3.0.9
distro==1.9.0
exceptiongroup==1.2.0
fairseq==0.12.2
fastapi==0.110.0
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hydra-core==1.0.7
idna==3.6
jieba==0.42.1
Jinja2==3.1.3
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.0.6
openai==1.13.3
packaging==23.2
portalocker==2.8.2
protobuf==4.25.3
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
sacrebleu==2.4.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
torch==2.2.1
torchaudio==2.2.1
tqdm==4.66.2
typing_extensions==4.10.0
uvicorn==0.27.1
